返回记录运算

+++ Api/src/main/java/com/java110/api/listener/floor/QueryFloorsListener.java
@@ -69,7 +69,7 @@ public class QueryFloorsListener extends AbstractServiceApiDataFlowListener {
             apiFloorVo.setApiFloorDataVoList(BeanConvertUtil.covertBeanList(floorDtoList, ApiFloorDataVo.class));
         }
 
-        apiFloorVo.setRecords((int) Math.ceil(total / row));
+        apiFloorVo.setRecords((int) Math.ceil((double) total / (double) row));
 
         ResponseEntity<String> responseEntity = new ResponseEntity<String>(JSONObject.toJSONString(apiFloorVo), HttpStatus.OK);
         dataFlowContext.setResponseEntity(responseEntity);
